changed amount form int to float

+++ Asel Salmenbayeva/GoAPI/types.go
@@ -2,6 +2,6 @@ package main
 
 type payment struct {
 	PaymentID   string
-	Amount     int       `bson:"Amount" json:"Amount"`
+	Amount     float32       `bson:"Amount" json:"Amount"`
 
 }
